requirements.txt
django>=3.0.3,<3.1.0
psycopg2>=2.7.5,<2.8.0
pillow>=6.0.0,<7.0.0
flake8>=3.6.0,<3.7.0
django-countries>=5.0.0,<7.0.0
django-phone-field
django-imagekit>=4.0.0,<4.0.5